Write setup script linux (#825)
* Mac setup created

structure of the setup for mac has been implemented, now it just needs debugging and testing to make sure it works.

* Script fixes

fixed some issues surround certain command within the mac setup not working. These should have been fixed now.

* added Documentation

added comments and documentation for the script and it's functions, along with some warnings about use.

* pep8

* Minor fix to yarn command

added a new way of doing the frontend dependencies part of the setup that should solve the unreliability issue.

* pep8++

* pep8++

* Structured linux setup script & minor changes to mac script

written the structure of the linux/ubuntu setup script. There are some additional checks that need to be implemented at some point, like checking if they're using something other than an AMD or intel cpu.

* Testing on Linux & minor-ish changes.

* Minor changes

finished the changes needed from the testing on linux.  The script now installs aimmo on ubuntu, in future this may needed to be expanded to cover other linux distributions. But for now this is fine.

* Minor changes++

* pep8

* pep8++ with small refactor

code has been refactored to hopefully make it easier to understand and to resolve pep8 issues.

* grammar fix in prints

* Major refactor

After some discussion and some pep8 related warnings, i went for a major refactor of the script in terms of it's structure to make it less complex, and hopefully further improve readability.

* More refactor changes

* pep8

* Minor changes

* Minor fix

fixed nodejs not always working

* Fixed dpkg-query not working

* Fixed get_nodejs()

* Removed redundant code and minor fixes

* minor changes

* Minor changes + spelling corrections.

* Whitespace + grammar