add logic for buttons error state (read desc.)
This would occur when the frontend cannot communicate with the server, the way this is usually handled is via a timeout. I've added an epic & reducer that will set a timeoutStatus to true after 20 seconds of inactivity from the server (there's an error though). The game should "un-timeout" when it receives a request, this still needs to be done, along with the visuals of the button and the popup message.