Run code button states (#927)
* add updating state

* remove debug console log

* Merge branch 'development' into run_code_loading

* circular progress inherits it’s coor from the parent button

* disable button when updating

* implementation of runcodebutton states (tests failing)

* fix proptypes not working

* fix/add tests

* add test case

* add logic for buttons error state (read desc.)

This would occur when the frontend cannot communicate with the server, the way this is usually handled is via a timeout. I've added an epic & reducer that will set a timeoutStatus to true after 20 seconds of inactivity from the server (there's an error though). The game should "un-timeout" when it receives a request, this still needs to be done, along with the visuals of the button and the popup message.

* moved timeoutStatus into game

* timeout state changes to true when no response received after x seconds

* Simplified approach to just detect when gameState is not being received

* cleanup

* map timeoutStatus state to button prop

* cleanup++

* tests for error state

* bump version

* fix failing tests